CORE: add v5.20 templates
and fixup (symlink) v5.14 template

+++ t/CORE-v5.14/C-COMPILED/template.pl
@@ -5,42 +5,64 @@ use warnings;
 
 use TAP::Harness ();
 use IO::Scalar;
-
+use Cwd;
+use File::Basename;
 use Test::More;
 
-#my @optimizations = ( '-O2,-fno-fold', '-O1' );
-my @optimizations = $ENV{BC_OPT} ? split(/\s+/,$ENV{BC_OPT}) : ('-O0','-O3');
-my $todo       = '';
+BEGIN {
+    use FindBin;
+    unshift @INC, $FindBin::Bin . "/../../../lib";
+}
+
+use KnownErrors qw/check_todo/;
+
+if ( $0 =~ m{/template\.pl$} ) {
+    plan q{skip_all} => "This program is not designed to be called directly";
+    exit;
+}
+
+my @optimizations = $ENV{'BC_TEST_OPTIMIZATIONS'} || '-O3,-fno-fold';
+$optimizations[0] .= ',-v'     if ( $ENV{VERBOSE} );
+$optimizations[0] .= ',-fwalkall' if ( $ENV{BC_WALK} );
 
 # Setup file_to_test to be the file we actually want to test.
-my $file_to_test = $0;
-if ( $file_to_test =~ s{==(.*)\.t$}{.t} ) {
-    my $options = $1;
-    $todo = "B::C Fails to generate c code. Issues: $1"         if ( $options =~ /BC-([\d-]+)/ );
-    $todo = "gcc cannot compile generated c code. Issues: $1"   if ( $options =~ /GCC-([\d-]+)/ );
-    $todo = "Compiled binary exits with signal. Issues: $1"     if ( $options =~ /SIG-([\d-]+)/ );
-    $todo = "Test crashes before completion. Issues: $1"        if ( $options =~ /BADPLAN-([\d-]+)/ );
-    $todo = "Fails tests when compiled with perlcc. Issues: $1" if ( $options =~ /BADTEST-([\d-]+)/ );
-    $todo = "Tests out of sequence. Issues: $1"                 if ( $options =~ /SEQ-([\d-]+)/ );
-    $todo = "TODO test unexpectedly passing. Issues: $1"        if ( $options =~ /TODO-([\d-]+)/ );
+my ( $file_to_test, $path ) = fileparse($0);
+my ( $before, $after ) = split( 'C-COMPILED/', $path );
+my $short_path = $path;
+$short_path =~ s{^.*C-COMPILED/+}{};
+$file_to_test = $short_path . $file_to_test;
+
+# The file that tracks acceptable failures in the compiled unit tests.
+my $known_errors_file = "known_errors.txt";
+
+my $errors = KnownErrors->new( file_to_test => $file_to_test );
+
+# The relative path our symlinks will point to.
+my $base_dir = dirname($path);
+
+chdir "$FindBin::Bin/../" or die $!;
+
+my $current_t_file   = $file_to_test;
+my $todo_description = '';
+
+my $type = $errors->get_current_error_type();
+$todo_description = $errors->{todo_description} if $type;
+
+# Skip this test all together if $type is SKIP or COMPAT
+if ( $type eq 'COMPAT' || $type eq 'SKIP' ) {
+    plan skip_all => $todo_description;
 }
 
-$file_to_test =~ s{--}{/}g;
-$file_to_test =~ s{C-COMPILED/}{};    # Strip the BINARY dir off to look for this test elsewhere.
+# need to run CORE test suite in t
+chdir "$FindBin::Bin/../../t" or die "Cannot chdir to t directory: $!";
 
-if ( $] < 5.014 && $file_to_test =~ m{^t/CORE/} ) {
-    plan skip_all => "Perl CORE tests only supported since 5.14 right now.";
-}
-else {
-    plan tests => 3 + 10 * scalar @optimizations;
-}
+plan tests => 3 + 9 * scalar @optimizations;
 
 ok( !-z $file_to_test, "$file_to_test exists" );
-
 open( my $fh, '<', $file_to_test ) or die("Can't open $file_to_test");
 my $taint = <$fh>;
 close $fh;
-$taint = ( ( $taint =~ m/\s\-T/ ) ? '-T' : '' );
+$taint = ( ( $taint =~ m/\s\-w?T/ ) ? '-T' : '' );
 pass( $taint ? "Taint mode!" : "Not in taint mode" );
 
 ( my $c_file   = $file_to_test ) =~ s/\.t$/.c/;
@@ -48,6 +70,7 @@ pass( $taint ? "Taint mode!" : "Not in taint mode" );
 unlink $bin_file, $c_file;
 
 my $PERL = $^X =~ m/\s/ ? qq{"$^X"} : $^X;
+my $blib = ( grep { m{blib/} } @INC ) ? '-I../../../blib/arch -I../../../blib/lib' : '';
 
 my $check = `$PERL -c $taint '$file_to_test' 2>&1`;
 like( $check, qr/syntax OK/, "$PERL -c $taint $file_to_test" );
@@ -58,50 +81,37 @@ my %SIGNALS = qw( 11 SEGV 6 SIGABRT 1 SIGHUP 13 SIGPIPE);
 $SIGNALS{0} = '';
 
 foreach my $optimization (@optimizations) {
-TODO: {
-  SKIP: {
-        local $TODO = $todo if ( $todo =~ /B::C Fails to generate c code/ );
-        local $ENV{BC_OPT} = $optimization;
+  TODO: SKIP: {
+        local $TODO;
 
-        my $b = $optimization; # protect against parallel test name clashes
-        #$b =~ s/-(D.*|f.*|v),//g;
-        #$b =~ s/-/_/g;
-        #$b =~ s/[, ]//g;
-        #$b =~ s/_O0$//;
-        #$b = lc($b);
-        $b = ''; # need to check $0 diagnostics
-        ( $c_file   = $file_to_test ) =~ s/\.t$/$b.c/;
-        $b = '.bin'; # need to check $0 diagnostics
-        ( $bin_file = $file_to_test ) =~ s/\.t$/$b/;
-        unlink $bin_file, $c_file;
+        $errors->{to_skip} = 9;
 
         # Generate the C code at $optimization level
-        my $cmd = "$PERL $taint -Iblib/arch -Iblib/lib -MO=-qq,C,$optimization,-o$c_file $file_to_test 2>&1";
+        my $cmd = "$PERL $blib $taint -MO=-qq,C,$optimization,-o$c_file $file_to_test 2>&1";
 
-        diag $cmd if $ENV{TEST_VERBOSE};
+        diag '$ '.getcwd() if $ENV{VERBOSE};
+        diag $cmd if $ENV{VERBOSE};
         my $BC_output = `$cmd`;
-        note $BC_output if ($BC_output);
-        ok( !-z $c_file, "$c_file is generated ($optimization)" );
+        note $BC_output if $BC_output;
 
-        if ( -z $c_file ) {
-            unlink $c_file;
-            skip( "Can't test further due to failure to create a c file.", 9 );
+        unless ( $errors->check_todo( -e $c_file && !-z _, "$c_file is generated ($optimization)", 'BC' ) ) {
+            unlink $c_file unless $ENV{BC_DEVELOPING};
+            skip( "Can't test further due to failure to create a c file.", $errors->{to_skip} );
         }
 
         # gcc the c code.
-        local $TODO = $todo if ( $todo =~ /gcc cannot compile generated c code/ );
-
-        $cmd = "$PERL -Iblib/arch -Iblib/lib script/cc_harness -q $c_file -o $bin_file 2>&1";
-        diag $cmd if $ENV{TEST_VERBOSE};
+        my $harness_opts = '';
+        $harness_opts = '-Wall' if $ENV{VERBOSE} && $ENV{WARNINGS};
+        $harness_opts .= $ENV{VERBOSE} ? '' : ' -q';
+        $cmd = "$PERL $FindBin::Bin/../../../../script/cc_harness $harness_opts $c_file -o $bin_file 2>&1";
+        diag $cmd if $ENV{VERBOSE};
         my $compile_output = `$cmd`;
-        note $compile_output if $compile_output;
+        note $compile_output if ($compile_output);
 
         # Validate compiles
-        ok( -x $bin_file, "$bin_file is compiled and ready to run." );
-
-        if ( !-x $bin_file ) {
+        unless ( $errors->check_todo( -x $bin_file, "$bin_file is compiled and ready to run.", 'GCC' ) ) {
             unlink $c_file, $bin_file unless $ENV{BC_DEVELOPING};
-            skip( "Can't test further due to failure to create a binary file.", 8 );
+            skip( "Can't test further due to failure to create a binary file.", $errors->{to_skip} );
         }
 
         # Parse through TAP::Harness
@@ -119,58 +129,42 @@ TODO: {
         close $out_fh;
 
         my $parser = $res->{parser_for}->{$bin_file};
-        ok( $parser, "Output parsed by TAP::Harness" );
+        $errors->check_todo( $parser, "Output parsed by TAP::Harness" ) or do {
+            skip "Cannot parse TAP output", $errors->{to_skip};
+        };
 
         my $signal = $res->{wait} % 256;
-        if ( $todo =~ /Compiled binary exits with signal/ ) {
-            local $TODO = "Tests don't pass at the moment - $todo";
-            my $sig_name = $SIGNALS{$signal};
-            ok( $signal == 0, "Exit signal is $signal ($sig_name)" );
-            note $out if ($out);
-            skip( "Test failures irrelevant if exits premature with $sig_name", 6 );
-        }
-        else {
-            ok( $signal == 0, "Exit signal is $signal" );
+        my $sig_name = $SIGNALS{$signal} || '';
+
+        unless ( $errors->check_todo( $signal == 0, "Exit signal is $signal $sig_name", 'SIG' ) ) {
+            note $out if $out;
+            skip( "Test failures irrelevant if exits premature with $sig_name", $errors->{to_skip} );
         }
 
-        if ( $todo =~ m/Test crashes before completion/ ) {
-            local $TODO = $todo;
-            ok( $parser->{is_good_plan}, "Plan was valid" );
-            note $out;
-            skip( "TAP parse is unpredictable when plan is invalid", 5 );
-        }
-        else {
-            ok( $parser->{is_good_plan}, "Plan was valid" );
+        unless ( $errors->check_todo( $parser->{is_good_plan}, "Plan was valid", 'PLAN' ) ) {
+            note $out if $out;
+            skip( "TAP parse is unpredictable when plan is invalid", $errors->{to_skip} );
         }
 
-        ok( $parser->{exit} == 0, "Exit code is $parser->{exit}" );
-
-        local $TODO = "Tests don't pass at the moment - $todo"
-          if ( $todo =~ /Fails tests when compiled with perlcc/ );
-        ok( !scalar @{ $parser->{failed} }, "Test results:" );
+        my $tests_ok = $errors->check_todo( $parser->{exit} == 0, "Exit code is $parser->{exit}", "EXIT" );
+        $tests_ok = $errors->check_todo( !scalar @{ $parser->{failed} }, "Test results:", 'TESTS' ) && $tests_ok;
         print "    $_\n" foreach ( split( "\n", $out ) );
 
-        if (!ok( !scalar @{ $parser->{failed} }, "No test failures $optimization" )) {
-          note( "Failed $optimization tests: " . join( ", ", @{ $parser->{failed} } ) );
-          $ENV{BC_DEVELOPING} = 1; # keep temp files
+        unless ($tests_ok) {
+            note( "Failed tests: " . join( ", ", @{ $parser->{failed} } ) );
+            skip "tests are failing", $errors->{to_skip};
         }
 
-        skip( "Don't care about test sequence if tests are failing", 2 )
-          if ( $todo =~ /Fails tests when compiled with perlcc/ );
+        $errors->check_todo( !scalar @{ $parser->{parse_errors} }, "Tests are in sequence", 'SEQ' )
+          or do {
+            note explain $parser->{parse_errors};
+            skip "tests are not in sequence", $errors->{to_skip};
+          };
 
-        local $TODO = $todo if ( $todo =~ m/Tests out of sequence/ );
-        if (!ok( !scalar @{ $parser->{parse_errors} }, "Tests are in sequence" )) {
-          note explain $parser->{parse_errors};
-          $ENV{BC_DEVELOPING} = 1; # keep temp files
-        }
-
-        local $TODO = "tests unexpectedly passing" if scalar @{ $parser->{todo_passed} };
-        if (!ok( !scalar @{ $parser->{todo_passed} }, "No TODO tests passed $optimization" )) {
-          note( "TODO Passed: " . join( ", ", @{ $parser->{todo_passed} } ) );
-          $ENV{BC_DEVELOPING} = 1; # keep temp files
-        }
-        $TODO = '';
+        $errors->check_todo( !scalar @{ $parser->{todo_passed} }, "No TODO tests passed", 'TODO' )
+          or note( "TODO Passed: " . join( ", ", @{ $parser->{todo_passed} } ) );
     }
-  }
-  unlink $bin_file, $c_file unless $ENV{BC_DEVELOPING};
 }
+unlink $bin_file, $c_file unless $ENV{BC_DEVELOPING};
+
+exit;
